Bring braintree module nesting in line with go_cardless
