+++ app/models/member.rb
@@ -1,5 +1,5 @@
 class Member < ActiveRecord::Base
-  has_one :customer, class_name: "Payment::BraintreeCustomer"
+  has_one :customer,               class_name: "Payment::Braintree::Customer"
   has_many :go_cardless_customers, class_name: "Payment::GoCardless::Customer"
   has_paper_trail on: [:update, :destroy]
