--- lib/payment_processor/clients/braintree/merchant_account_selector.rb
@@ -1,19 +1,0 @@
-module PaymentProcessor
-  module Clients
-    module Braintree
-      class MerchantAccountSelector
-        include ActsLikeSelectorWithCurrency
-
-        MERCHANT_ACCOUNTS = Settings.braintree.merchants.freeze
-
-        def select_or_raise
-          raise_error if @currency.blank?
-          id = MERCHANT_ACCOUNTS[@currency.upcase.to_sym]
-          raise_error("No merchant account is associated with this currency: #{@currency}")  unless id
-          id
-        end
-      end
-    end
-  end
-end
-
