+++ spec/factories/braintree_payment_method.rb
@@ -1,5 +1,5 @@
 FactoryGirl.define do
-  factory :braintree_payment_method, :class => 'Payment::BraintreePaymentMethod' do
+  factory :braintree_payment_method, :class => 'Payment::Braintree::PaymentMethod' do
     customer_id { Faker::Number.number(6) }
     token "MyString"
   end
