+++ spec/factories/payment_braintree_subscriptions.rb
@@ -1,5 +1,5 @@
 FactoryGirl.define do
-  factory :payment_braintree_subscription, class: 'Payment::BraintreeSubscription' do
+  factory :payment_braintree_subscription, class: 'Payment::Braintree::Subscription' do
     subscription_id { "s#{Faker::Number.number(4)}" }
     amount 79.41
     currency "GBP"
