+++ spec/factories/payment_braintree_transactions.rb
@@ -1,5 +1,5 @@
 FactoryGirl.define do
-  factory :payment_braintree_transaction, :class => 'Payment::BraintreeTransaction' do
+  factory :payment_braintree_transaction, :class => 'Payment::Braintree::Transaction' do
     transaction_id { "t#{Faker::Number.number(4)}" }
     transaction_type "credit_card"
     status 0
