--- spec/lib/payment_processor/clients/braintree/subscription_plan_selector_spec.rb
@@ -1,32 +1,0 @@
-require 'rails_helper'
-
-module PaymentProcessor
-  module Clients
-    module Braintree
-      describe SubscriptionPlanSelector do
-        before do
-          stub_const("PaymentProcessor::Clients::Braintree::SubscriptionPlanSelector::SUBSCRIPTION_PLANS", { EUR: 'EUR' })
-        end
-
-        subject { described_class }
-
-        describe '.for_currency' do
-          context 'unmatched currency' do
-            it 'raises' do
-              expect{
-                subject.for_currency('VVZ')
-              }.to raise_error(Exceptions::InvalidCurrency, 'No merchant account is associated with this currency: VVZ')
-            end
-          end
-
-          context 'matched currency' do
-            it 'returns merchant account ID' do
-              expect( subject.for_currency('EUR') ).to eq('EUR')
-            end
-          end
-        end
-      end
-    end
-  end
-end
-
