+++ spec/models/payment/braintree/subscription_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-describe Payment::BraintreeSubscription do
+describe Payment::Braintree::Subscription do
 
   let(:subscription) { create :payment_braintree_subscription }
   subject { subscription }
@@ -23,8 +23,7 @@ describe Payment::BraintreeSubscription do
   it 'handles money properly' do
     create :payment_braintree_subscription, amount: 12.41
     create :payment_braintree_subscription, amount: 10701.11
-    expect(Payment::BraintreeSubscription.all.map(&:amount).sum).to eq 10713.52
-    expect(Payment::BraintreeSubscription.last.amount.class).to eq BigDecimal
+    expect(Payment::Braintree::Subscription.all.map(&:amount).sum).to eq 10713.52
+    expect(Payment::Braintree::Subscription.last.amount.class).to eq BigDecimal
   end
-
 end
